+++ src/analysis/retail/evoker/devastation/Guide.tsx
@@ -4,7 +4,7 @@ import PreparationSection from 'interface/guide/components/Preparation/Preparati
 import { CooldownSection } from './modules/guide/Cooldown';
 import { CoreRotation } from './modules/guide/CoreRotation';
 import { DamageEfficiency } from './modules/guide/DamageEfficiencySection';
-import { EssenceGraphSection } from './modules/guide/EssenceGraphSection';
+import { CoreSection } from './modules/guide/CoreSection';
 import { DragonRageSection } from './modules/guide/DragonRageSection';
 import { IntroSection } from './modules/guide/IntroSection';
 import MajorDefensives from '../shared/modules/MajorDefensives/DefensivesGuide';
@@ -13,7 +13,7 @@ export default function Guide({ modules, events, info }: GuideProps<typeof Comba
   return (
     <>
       <IntroSection />
-      <EssenceGraphSection modules={modules} info={info} events={events} />
+      <CoreSection modules={modules} info={info} events={events} />
       <CooldownSection modules={modules} info={info} events={events} />
       <DragonRageSection modules={modules} info={info} events={events} />
       <DamageEfficiency modules={modules} info={info} events={events} />
